Fix VHPI cast issue #2126
vhpiIntT before the 2008 standard was defined as an unsigned integer in
vhpi_user.h. However, the standard states that all integers are 32-bit
integers with two's complement representation.

The issue occurs when vhpiIntT is defined as uint32_t and long is a
signed 64 bit integer. Negative values are represented as large unsigned
values in two's complement. When this value is returned as a long, it is
implicitly converted ("safely") to the larger signed representation.

To prevent this issue we convert the vhpiIntT into a 32-bit two's
complement integer type (int32_t) before allowing the conversion to
take place.

+++ cocotb/share/lib/vhpi/VhpiCbHdl.cpp
@@ -789,7 +789,7 @@ long VhpiSignalObjHdl::get_signal_value_long()
         LOG_ERROR("VHPI: Failed to get value of type long");
     }
 
-    return value.value.intg;
+    return static_cast<int32_t>(value.value.intg);
 }
